asdf

+++ setup.sh
@@ -4,24 +4,24 @@
 git config --global user.email "sunladenmail@gmail.com"
 git config --global user.name "sunladen"
 
-# configure credential.helper if in a WSL environment and Git for Windows git-credential-manager can be found
+# configure credential.helper
 GCM_REL_EXE="mingw64/libexec/git-core/git-credential-manager-core.exe"
 if grep -qi microsoft /proc/version; then
+	# in a WSL environment
 	USERPROFILE=$(wslpath "$(wslvar USERPROFILE)")
 	PROGFILES_PATH="/mnt/c/Program\ Files/Git"
 elif command -v cygpath &> /dev/null; then
+	# in a Git for Windows environment
+	USERPROFILE="${HOME}"
 	PROGFILES_PATH="/c/Program\ Files/Git"
 fi
 
 if [ -f "${USERPROFILE}/AppData/Local/Programs/Git/${GCM_REL_EXE}" ]; then
-	echo "GCM in local user path"
-	#git config --global credential.helper "/mnt/c/Program\ Files/Git/${GCM_EXE}"
-elif [ -f "${PROGFILES_PATH}/${GCM_REL_EXE}" ]; then
-	echo "GCM in Program Files path"
-	#git config --global credential.helper "/mnt/c/Program\ Files/Git/${GCM_EXE}"
+	git config --global credential.helper "${USERPROFILE}/AppData/Local/Programs/Git/${GCM_REL_EXE}"
+elif [ -f "${PROGFILES_PATH}$/{GCM_REL_EXE}" ]; then
+	git config --global credential.helper "${PROGFILES_PATH}/${GCM_REL_EXE}"
 fi
 
-exit 0
 
 
 # disable login banner
@@ -31,13 +31,16 @@ touch ~/.hushlogin
 # make sure there is a $HOME/.config directory
 mkdir -p ~/.config
 
+# make sure there is a $HOME/.vim directory
+mkdir -p ~/.vim
+
 
 # symlink .dotfiles
 rm -rf ~/.bash_aliases && ln -s $PWD/bash/.bash_aliases ~/.bash_aliases
 rm -rf ~/.bash_profile && ln -s $PWD/bash/.bash_profile ~/.bash_profile
 rm -rf ~/.bashrc && ln -s $PWD/bash/.bashrc ~/.bashrc
 rm -rf ~/.inputrc && ln -s $PWD/bash/.inputrc ~/.inputrc
-rm -rf ~/.vim && ln -s $PWD/.vim ~/.vim
+rm -rf ~/.vim/vimrc && ln -s $PWD/.vim/vimrc ~/.vim/vimrc
 rm -rf ~/.config/nvim && ln -s $PWD/.config/nvim ~/.config/nvim
 rm -rf ~/.tmux.conf && ln -s $PWD/.tmux.conf ~/.tmux.conf
 rm -rf ~/.config/ranger && ln -s $PWD/.config/ranger ~/.config/ranger
@@ -60,42 +63,42 @@ if [ ! -d $TPM_DIR ]; then
 fi
 
 
-# question whether to go on and perform package management and tool installs
-read -r -p "run package purge, install and updates (requires sudo)? [Yn] " response
-case "$response" in
-	[Yy])
-		sudo apt purge --auto-remove -y tmux
-		sudo apt purge --auto-remove -y nano
-		sudo apt purge --auto-remove -y vim-common
 
-		# Full apt update, upgrade, remove, clean cycle
-		sudo apt update
-		sudo apt upgrade -y
-		sudo apt dist-upgrade -y
-		sudo apt autoremove -y
-		sudo apt autoclean -y
+if grep -qi microsoft /proc/version; then
+	# in a WSL environment
 
-		# List installed packages
-		#sudo apt list --installed
+	# question whether to go on and perform package management and tool installs
+	read -r -p "run package purge, install and updates (requires sudo)? [Yn] " response
+	case "$response" in
+		[Yy])
+			sudo apt purge --auto-remove -y tmux
+			sudo apt purge --auto-remove -y nano
+			sudo apt purge --auto-remove -y vim-common
 
-		# Read size of WSL distro excluding Windows drive mount
-		#sudo du -sh / --exclude=/mnt/c
+			# Full apt update, upgrade, remove, clean cycle
+			sudo apt update
+			sudo apt upgrade -y
+			sudo apt dist-upgrade -y
+			sudo apt autoremove -y
+			sudo apt autoclean -y
 
-		sudo apt install -y wsl
-		sudo apt install -y fzf
-		sudo apt install -y -o Dpkg::Options::="--force-overwrite" bat ripgrep
+			# List installed packages
+			#sudo apt list --installed
 
-		sudo ./update.sh
-		;;
-esac
+			# Read size of WSL distro excluding Windows drive mount
+			#sudo du -sh / --exclude=/mnt/c
 
+			sudo apt install -y wsl
+			sudo apt install -y fzf
+			sudo apt install -y -o Dpkg::Options::="--force-overwrite" bat ripgrep
+			sudo apt install -y python3-pip
+
+			sudo ./update.sh
+			;;
+	esac
+
+fi
 
 
 # source new symlinked $HOME/.bashrc
 . ~/.bashrc
-
-# install pip
-sudo apt install python3-pip
-
-# user-level install powerline-status
-pip install --user powerline-status
